requirements.txt
jupyterlab
pandas
numpy
matplotlib
seaborn
scikit-learn
pyyaml
nbconvert
statsmodels
plotly
requests
